% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{bp_cat}
\alias{bp_cat}
\title{Blood presure categorization according to REGICOR model}
\usage{
bp_cat(dbp, sbp)
}
\arguments{
\item{dbp}{diastolic blood pressure}

\item{sbp}{systolic blood pressure}
}
\value{
BP levels
}
\description{
Blood presure categorization according to REGICOR model
}

