% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{ckd.epi}
\alias{ckd.epi}
\title{Gromerular filtration rate function}
\usage{
ckd.epi(cre, age, sex, race = 0)
}
\arguments{
\item{cre}{Creatinine level}

\item{age}{Age}

\item{sex}{sex}

\item{race}{0 = white, 1 = black}
}
\value{
Gromerular filtration rate function
}
\description{
Gromerular filtration rate function
}

