% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{ckd.stage}
\alias{ckd.stage}
\title{Gromerular filtration rate function}
\usage{
ckd.stage(gfr)
}
\arguments{
\item{gfr}{Glomerular filtration rate}
}
\value{
Chronic kidney disease stage
}
\description{
Gromerular filtration rate function
}

