% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidences.R
\name{coef.incidence}
\alias{coef.incidence}
\title{Incidence coefficient}
\usage{
\method{coef}{incidence}(inc)
}
\arguments{
\item{inc}{\code{\link{incidence}} object}
}
\value{
coefficient for the incidence
}
\description{
Incidence coefficient
}

