% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{colhdl_cat}
\alias{colhdl_cat}
\title{High density cholesterol categorization according to REGICOR model}
\usage{
colhdl_cat(colhdl)
}
\arguments{
\item{colhdl}{High density cholesterol}
}
\value{
colhdl levels
}
\description{
High density cholesterol categorization according to REGICOR model
}

