% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_create.R
\name{coltot_cat}
\alias{coltot_cat}
\title{Total cholesterol categorization according to REGICOR model}
\usage{
coltot_cat(coltot)
}
\arguments{
\item{coltot}{total cholesterol}
}
\value{
coltot levels
}
\description{
Total cholesterol categorization according to REGICOR model
}

