% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_create.R
\name{create.endpoints}
\alias{create.endpoints}
\title{This function converts a given outcome in date format with time and event variables}
\usage{
create.endpoints(dataset, time_to_event = TRUE, check_exitus = FALSE,
  dexitus = "dexitus", eps = names(dataset)[substring(names(dataset), 1, 3)
  == "ep_"], silently = TRUE)
}
\arguments{
\item{dataset}{original dataset}

\item{time_to_event}{calculate time to event}

\item{check_exitus}{if TRUE all eps variables have to appear after dexitus date}

\item{dexitus}{(default: 'dexitus')}

\item{eps}{endpoints to calculate. (default: all variables starting with 'ep_')}

\item{silently}{print messages (default = FALSE)}
}
\value{
original dataset with time and event variables included
}
\description{
This function converts a given outcome in date format with time and event variables
}

