% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_create.R
\name{create.ep_combinate}
\alias{create.ep_combinate}
\title{Create a combined end-point from different endpoints}
\usage{
create.ep_combinate(deps)
}
\arguments{
\item{deps}{enpoints}
}
\value{
returns the combined endpoint
}
\description{
Create a combined end-point from different endpoints
}

