% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_create.R
\name{create.ep_death}
\alias{create.ep_death}
\title{Calculate an date event variable for death}
\usage{
create.ep_death(exitus, dexitus, death = "D")
}
\arguments{
\item{exitus}{exitus reason}

\item{death}{level indicating death}

\item{date}{of exitus}
}
\value{
creates a date variable with the date of death for dose dying
}
\description{
Calculate an date event variable for death
}

