% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{create.statine_level}
\alias{create.statine_level}
\title{Fix the level of statine treatment, using PETREA definition or binary}
\usage{
create.statine_level(dose, atc, definition = "PETREA")
}
\arguments{
\item{dose}{statin dose}

\item{atc}{statin type}

\item{definition}{character 'PETREA' (default), '2-levels'}
}
\value{
returns the statin level
}
\description{
Fix the level of statine treatment, using PETREA definition or binary
}

