% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc.R
\name{cuminc.mi}
\alias{cuminc.mi}
\title{Build a dataframe with the Nelson-Aalen survival estimates}
\usage{
cuminc.mi(mi_data, str_coxph = "Surv(time = time, event = event)~user",
  min_obs = 5)
}
\arguments{
\item{mi_data}{multiple imputation data with each imputation in stacked by rows}

\item{str_coxph}{string containing the cox models to be evaluated (default="Surv(time = time, event = event)~user")}

\item{min_obs}{minimum number of imputations with event in specific time to evaluated the average}
}
\value{
result table with survival time
}
\description{
Build a dataframe with the Nelson-Aalen survival estimates
}

