% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{filter.bp}
\alias{filter.bp}
\title{Filter bp variable eliminating measurements below and above a constant inf and sup respectively and with coherent measurement
between sbp and dbp}
\usage{
filter.bp(sbp, dbp, inf.sbp = 60, inf.dbp = 30, sup.sbp = 250,
  sup.dbp = 170)
}
\arguments{
\item{sbp}{vector with bmi measurements to be filtered}

\item{dbp}{indicative variable of obesity}

\item{sbp.inf}{lower bound (default 60)}

\item{dbp.inf}{lower bound (default 30)}

\item{sbp.sup}{upper bound (default 250)}

\item{dbp.sup}{upper bound (default 170)}
}
\value{
the measurement vector filtered with NA in measurements outside the rank [ind, sup]
}
\description{
Filter bp variable eliminating measurements below and above a constant inf and sup respectively and with coherent measurement
between sbp and dbp
}

