% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{filter.cholesterol}
\alias{filter.cholesterol}
\title{Filter liidic variables (coltot, colhdl, colldl, tg)}
\usage{
filter.cholesterol(coltot, colhdl, colldl, tg, inf.tot = 90, sup.tot = 450,
  inf.hdl = 25, sup.hdl = 150, inf.ldl = 50, sup.ldl = 295,
  inf.tg = 25, sup.tg = 2000)
}
\arguments{
\item{coltot}{coltot variable}

\item{colhdl}{colhdl variable}

\item{colldl}{colldl variable}

\item{tg}{tg variable}

\item{inf.tot}{coltot inferior limit}

\item{sup.tot}{coltot superior limit}

\item{inf.hdl}{colhdl inferior limit}

\item{sup.hdl}{colhdl superior limit}

\item{inf.ldl}{colldl inferior limit}

\item{sup.ldl}{colldl superior limit}

\item{inf.tg}{tg inferior limit}

\item{sup.tg}{tg superior limit}
}
\value{
coltot,colhdl,colldl and tg filtered
}
\description{
Filter liidic variables (coltot, colhdl, colldl, tg)
}

