% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{filter.variable}
\alias{filter.variable}
\title{Filter a variable eliminating measurements below and above a constant inf and sup respectively}
\usage{
filter.variable(value, inf, sup)
}
\arguments{
\item{value}{vector with measurements to be filtered}

\item{inf}{lower bound}

\item{sup}{upper bound}
}
\value{
the measurement vector filtered with NA in measurements outside the rank [ind, sup]
}
\description{
Filter a variable eliminating measurements below and above a constant inf and sup respectively
}

