% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{format.variables}
\alias{format.variables}
\title{Fix the level of statine treatment, using PETREA definition or binary}
\usage{
\method{format}{variables}(dataset, vnumeric = c(), vdate = c(),
  vfactor = c(), silently = FALSE)
}
\arguments{
\item{silently}{}
}
\value{
returns dataset with variables converted
}
\description{
Fix the level of statine treatment, using PETREA definition or binary
}

