% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_create.R
\name{format_variables}
\alias{format_variables}
\title{Format variables}
\usage{
format_variables(dataset, vnumeric = c(), vdate = c(), vfactor = c(),
  silently = FALSE)
}
\arguments{
\item{silently}{}
}
\value{
returns dataset with variables converted
}
\description{
Format variables
}

