% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{get_icd}
\alias{get_icd}
\title{Returns the icd code for a disease}
\usage{
get_icd(disease, f_xml, icd_list, icd = 10)
}
\arguments{
\item{disease}{name of disease}

\item{f_xml}{xml file with the definitions}

\item{icd_list}{}
}
\value{
coefficient for the incidence
}
\description{
Returns the icd code for a disease
}

