% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ifelse.Date}
\alias{ifelse.Date}
\title{fitting a cox model to a multiple imputation dataset}
\usage{
ifelse.Date(cond, yes, no)
}
\arguments{
\item{cond}{condition}

\item{yes}{result when TrUE}

\item{no}{result when FALSE}
}
\value{
combinatino of yes and no depending on condition
}
\description{
fitting a cox model to a multiple imputation dataset
}

