% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/incidences.R
\name{print.incidence}
\alias{print.incidence}
\title{Incidence summary}
\usage{
\method{print}{incidence}(x)
}
\arguments{
\item{inc}{\code{\link{incidence}} object}
}
\value{
print an incidence object
}
\description{
Incidence summary
}

