% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.alcoholism}
\alias{recover.alcoholism}
\title{Recover alcoholism}
\usage{
recover.alcoholism(alcoholism, alcohol, na.value = 0)
}
\arguments{
\item{alcoholism}{original indicator variable for smoking condition}

\item{alcohol}{alcohol condition (0: no drinker, 1: low risk drinker, 2: high risk drinker)}

\item{na.value}{value to be assigned to people with missing in alcoholism variable}
}
\value{
recovered indicator variable for alcoholism condition from high risk drinker
}
\description{
Recover alcoholism
}

