% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.bmi}
\alias{recover.bmi}
\title{Recover bmi variable from height and weight}
\usage{
recover.bmi(bmi, height, weight)
}
\arguments{
\item{bmi}{vector with bmi measurements to be filtered}

\item{height}{height in cm}

\item{weight}{weight in kg}
}
\value{
bmi with recovered measurements
}
\description{
Recover bmi variable from height and weight
}

