% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.colesterol}
\alias{recover.colesterol}
\title{Recover cholesterols}
\usage{
recover.colesterol(coltot, colhdl, colldl, tg, filter = TRUE)
}
\arguments{
\item{coltot}{}

\item{colhdl}{}

\item{colldl}{}

\item{tg}{}

\item{filter}{apply the filter after recovering (default TRUE)}
}
\value{
recovered cholesterols
}
\description{
Recover cholesterols
}

