% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.obesity}
\alias{recover.obesity}
\title{Recover obesity from BMI calculation}
\usage{
recover.obesity(obesity, imc)
}
\arguments{
\item{obesity}{original indicator variable for obesity condition}

\item{bmi}{vector with bmi measurements}
}
\value{
recovered indicator variable for obesity condition
}
\description{
Recover obesity from BMI calculation
}

