% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.smoking}
\alias{recover.smoking}
\title{Recover smoking}
\usage{
recover.smoking(smoking, smoke, t_smoke, non_smoker_time = 365,
  na.value = 0)
}
\arguments{
\item{smoking}{original indicator variable for smoking condition}

\item{smoke}{smoking condition (0: never, 1: smoker, 2: previous smoker)}

\item{t_smoke}{days since  smoking condition}

\item{non_smoker_time}{days since a person is considered non-smoker (default: 365)}

\item{na.value}{value to be assigned to people with missing in smoke variable}
}
\value{
recovered indicator variable for smoking condition
}
\description{
Recover smoking
}

