% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.using_medication}
\alias{recover.using_medication}
\title{Recover alcoholism}
\usage{
recover.using_medication(problem, meds)
}
\arguments{
\item{problem}{indicator variable for a given problem}

\item{meds}{matrix with indicator facturation associated to the problem}
}
\value{
recovered indicator variable for a given problem
}
\description{
Recover alcoholism
}

