% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/incidences.R
\name{vcov.incidence}
\alias{vcov.incidence}
\title{Incidence coefficient}
\usage{
\method{vcov}{incidence}(inc)
}
\arguments{
\item{inc}{\code{\link{incidence}} object}
}
\value{
variance estimator for the incidence
}
\description{
Incidence coefficient
}

