% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_recover-filter.R
\name{recover.cholesterol}
\alias{recover.cholesterol}
\title{Recover cholesterols (only if tg < 400)}
\usage{
recover.cholesterol(coltot, colhdl, colldl, tg, filter = TRUE)
}
\arguments{
\item{colldl}{}

\item{tg}{}

\item{filter}{apply the filter after recovering (default TRUE)}
}
\value{
recovered cholesterols
}
\description{
Recover cholesterols (only if tg < 400)
}

